Extract license plate helpers to dedicated module
